pyproject.toml
[tool.poetry]
name = "invoice-converter"
version = "0.1.0"
description = "A tool to update PDF invoice packages with new format invoices"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
Pillow = "^10.0.0"
pytesseract = "^0.3.10"
pdfplumber = "^0.9.0"
pdfrw = "^0.4.0"
tomli = "^2.0.1"  # For TOML parsing in Python <3.11
tomli-w = "^1.0.0"  # For writing TOML files
opencv-python-headless = "^4.8.0"  # For improved image processing
rapidfuzz = "^3.13.0"  # Modern and faster alternative to fuzzywuzzy + Levenshtein
camelot-py = "^0.11.0"  # For table extraction from PDFs

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^24.2.0"
isort = "^5.13.2"
flake8 = "^7.0.0"
mypy = "^1.8.0"
bandit = "^1.7.8"
pre-commit = "^3.5.0"
setuptools = "^78.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
invoice-converter = "invoice_converter:main"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.black]
line-length = 120
target-version = ["py312"]

[tool.isort]
profile = "black"
line_length = 120

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
package_root = "."
packages = ["invoice_converter"]

[tool.flake8]
max-line-length = 120
extend-ignore = "E203"
exclude = [".venv", "venv", "env", ".git", "build", "dist"]

[tool.invoice_converter.template]
# Define the single template that the system should recognize
template_path = "template.pdf"
# Fields to recognize with their expected coordinates [x1, y1, x2, y2] and types
fields = [
    {name = "invoice_number", type = "text", region = [0.7, 0.1, 0.9, 0.15]},
    {name = "date", type = "date", region = [0.7, 0.15, 0.9, 0.2]},
    {name = "customer_name", type = "text", region = [0.1, 0.2, 0.5, 0.25]},
    {name = "total_amount", type = "amount", region = [0.7, 0.8, 0.9, 0.85]}
]

[tool.invoice_converter.field_recognition]
# Configuration for field recognition
confidence_threshold = 0.85
use_fuzzy_matching = true
allow_multiple_matches = false
threshold_for_fuzzy_match = 90  # Percentage of similarity required for fuzzy matching
